Issue-2469 added support for http client builder
